Removing compiler warnings in order to avoid using -Wno-sign-compare -Wno-unused-value -Wno-unused-variable -Wno-unused-but-set-variable

+++ src/data.cpp
@@ -101,7 +101,7 @@ void InternDataHandler::convert(
     if (args_->trainMode == 1) {
       // pick one random label as rhs and the rest is lhs
       auto idx = rand() % example.RHSTokens.size();
-      for (int i = 0; i < example.RHSTokens.size(); i++) {
+      for (unsigned int i = 0; i < example.RHSTokens.size(); i++) {
         auto tok = example.RHSTokens[i];
         if (i == idx) {
           rslt.RHSTokens.push_back(tok);
@@ -113,7 +113,7 @@ void InternDataHandler::convert(
     if (args_->trainMode == 2) {
       // pick one random label as lhs and the rest is rhs
       auto idx = rand() % example.RHSTokens.size();
-      for (int i = 0; i < example.RHSTokens.size(); i++) {
+      for (unsigned int i = 0; i < example.RHSTokens.size(); i++) {
         auto tok = example.RHSTokens[i];
         if (i == idx) {
           rslt.LHSTokens.push_back(tok);
@@ -125,7 +125,7 @@ void InternDataHandler::convert(
     if (args_->trainMode == 3) {
       // pick two random labels, one as lhs and the other as rhs
       auto idx = rand() % example.RHSTokens.size();
-      int idx2;
+      unsigned int idx2;
       do {
         idx2 = rand() % example.RHSTokens.size();
       } while (idx2 == idx);
@@ -145,14 +145,14 @@ void InternDataHandler::getWordExamples(
     vector<ParseResults>& rslts) const {
 
   rslts.clear();
-  for (int widx = 0; widx < doc.size(); widx++) {
+  for (int widx = 0; widx < (int)(doc.size()); widx++) {
     ParseResults rslt;
     rslt.LHSTokens.clear();
     rslt.RHSTokens.clear();
     rslt.RHSTokens.push_back(doc[widx]);
-    for (int i = max(widx - args_->ws, 0);
+    for (unsigned int i = max(widx - args_->ws, 0);
          i < min(size_t(widx + args_->ws), doc.size()); i++) {
-      if (i != widx) {
+      if ((int)i != widx) {
         rslt.LHSTokens.push_back(doc[i]);
       }
     }
@@ -218,7 +218,7 @@ void InternDataHandler::getNextKExamples(int K, vector<ParseResults>& c) {
 void InternDataHandler::getRandomWord(vector<Base>& result) {
   result.push_back(word_negatives_[word_iter_]);
   word_iter_++;
-  if (word_iter_ >= word_negatives_.size()) {
+  if (word_iter_ >= (int)word_negatives_.size()) {
     word_iter_ = 0;
   }
 }
@@ -245,9 +245,9 @@ void InternDataHandler::getRandomRHS(vector<Base>& results) const {
   assert(size_ > 0);
   results.clear();
   auto& ex = examples_[rand() % size_];
-  int r = rand() % ex.RHSTokens.size();
+  unsigned int r = rand() % ex.RHSTokens.size();
   if (args_->trainMode == 2) {
-    for (int i = 0; i < ex.RHSTokens.size(); i++) {
+    for (unsigned int i = 0; i < ex.RHSTokens.size(); i++) {
       if (i != r) {
         results.push_back(ex.RHSTokens[i]);
       }
